+++ cumulus/parachains/integration-tests/emulated/chains/parachains/assets/asset-hub-westend/Cargo.toml
@@ -17,7 +17,6 @@ frame-support = { path = "../../../../../../../../substrate/frame/support", defa
 parachains-common = { path = "../../../../../../../parachains/common" }
 
 # Cumulus
-xcm-emulator = { path = "../../../../../../../xcm/xcm-emulator", default-features = false }
-integration-tests-common = { path = "../../../../common", default-features = false }
+emulated-integration-tests-common = { path = "../../../../common", default-features = false }
 asset-hub-westend-runtime = { path = "../../../../../../runtimes/assets/asset-hub-westend" }
 westend-emulated-chain = { path = "../../../relays/westend" }
