+++ cumulus/parachains/integration-tests/emulated/chains/parachains/assets/asset-hub-westend/src/lib.rs
@@ -19,7 +19,7 @@ pub mod genesis;
 use frame_support::traits::OnInitialize;
 
 // Cumulus
-use integration_tests_common::{
+use emulated_integration_tests_common::{
     xcm_emulator::decl_test_parachains,
     impl_accounts_helpers_for_parachain,
     impl_assert_events_helpers_for_parachain,
