+++ cumulus/parachains/integration-tests/emulated/chains/parachains/assets/asset-hub-wococo/src/lib.rs
@@ -19,7 +19,7 @@ pub mod genesis;
 use frame_support::traits::OnInitialize;
 
 // Cumulus
-use integration_tests_common::{
+use emulated_integration_tests_common::{
     xcm_emulator::decl_test_parachains,
     impl_accounts_helpers_for_parachain,
     impl_assert_events_helpers_for_parachain,
@@ -27,7 +27,7 @@ use integration_tests_common::{
 };
 use wococo_emulated_chain::Wococo;
 
-// AssetHubRococo Parachain declaration
+// AssetHubWococo Parachain declaration
 decl_test_parachains! {
 	pub struct AssetHubWococo {
 		genesis = genesis::genesis(),
@@ -52,7 +52,7 @@ decl_test_parachains! {
 	},
 }
 
-// AssetHubRococo implementation
+// AssetHubWococo implementation
 impl_accounts_helpers_for_parachain!(AssetHubWococo);
 impl_assert_events_helpers_for_parachain!(AssetHubWococo);
 impl_assets_helpers_for_system_parachain!(AssetHubWococo, Wococo);
