+++ cumulus/parachains/integration-tests/emulated/chains/parachains/bridges/bridge-hub-rococo/src/lib.rs
@@ -19,11 +19,10 @@ pub mod genesis;
 use frame_support::traits::OnInitialize;
 
 // Cumulus
-use integration_tests_common::{
+use emulated_integration_tests_common::{
     xcm_emulator::decl_test_parachains,
     impl_accounts_helpers_for_parachain,
     impl_assert_events_helpers_for_parachain,
-    // impl_assets_helpers_for_parachain,
 };
 
 // BridgeHubRococo Parachain declaration
@@ -50,4 +49,3 @@ decl_test_parachains! {
 // BridgeHubRococo implementation
 impl_accounts_helpers_for_parachain!(BridgeHubRococo);
 impl_assert_events_helpers_for_parachain!(BridgeHubRococo);
-// impl_assets_helpers_for_parachain!(AssetHub, Kusama);
