+++ cumulus/parachains/integration-tests/emulated/chains/parachains/bridges/bridge-hub-wococo/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.0.0"
 authors.workspace = true
 edition.workspace = true
 license = "Apache-2.0"
-description = "Bridge Hub Rococo emulated chain"
+description = "Bridge Hub Wococo emulated chain"
 publish = false
 
 [dependencies]
@@ -17,6 +17,5 @@ frame-support = { path = "../../../../../../../../substrate/frame/support", defa
 parachains-common = { path = "../../../../../../../parachains/common" }
 
 # Cumulus
-xcm-emulator = { path = "../../../../../../../xcm/xcm-emulator", default-features = false }
-integration-tests-common = { path = "../../../../common", default-features = false }
+emulated-integration-tests-common = { path = "../../../../common", default-features = false }
 bridge-hub-rococo-runtime = { path = "../../../../../../runtimes/bridge-hubs/bridge-hub-rococo" }
