+++ cumulus/parachains/integration-tests/emulated/chains/parachains/testing/penpal/Cargo.toml
@@ -17,6 +17,5 @@ frame-support = { path = "../../../../../../../../substrate/frame/support", defa
 parachains-common = { path = "../../../../../../../parachains/common" }
 
 # Cumulus
-xcm-emulator = { path = "../../../../../../../xcm/xcm-emulator" }
-integration-tests-common = { path = "../../../../common", default-features = false }
+emulated-integration-tests-common = { path = "../../../../common", default-features = false }
 penpal-runtime = { path = "../../../../../../runtimes/testing/penpal" }
