+++ cumulus/parachains/integration-tests/emulated/chains/relays/rococo/src/lib.rs
@@ -16,7 +16,7 @@
 pub mod genesis;
 
 // Cumulus
-use integration_tests_common::{
+use emulated_integration_tests_common::{
     xcm_emulator::decl_test_relay_chains,
     impl_accounts_helpers_for_relay_chain,
     impl_assert_events_helpers_for_relay_chain,
