+++ cumulus/parachains/integration-tests/emulated/chains/relays/wococo/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.0.0"
 authors.workspace = true
 edition.workspace = true
 license = "Apache-2.0"
-description = "Rococo emulated chain"
+description = "Wococo emulated chain"
 publish = false
 
 [dependencies]
@@ -15,18 +15,13 @@ sp-authority-discovery = { path = "../../../../../../../substrate/primitives/aut
 sp-consensus-babe = { path = "../../../../../../../substrate/primitives/consensus/babe", default-features = false }
 beefy-primitives = { package = "sp-consensus-beefy", path = "../../../../../../../substrate/primitives/consensus/beefy" }
 grandpa = { package = "sc-consensus-grandpa", path = "../../../../../../../substrate/client/consensus/grandpa", default-features = false }
-frame-support = { path = "../../../../../../../substrate/frame/support", default-features = false }
 pallet-im-online = { path = "../../../../../../../substrate/frame/im-online", default-features = false }
-pallet-staking = { path = "../../../../../../../substrate/frame/staking", default-features = false }
 
 # Polkadot
-polkadot-parachain-primitives = { path = "../../../../../../../polkadot/parachain", default-features = false }
 polkadot-primitives = { path = "../../../../../../../polkadot/primitives", default-features = false }
-polkadot-runtime-parachains = { path = "../../../../../../../polkadot/runtime/parachains", default-features = false }
 rococo-runtime-constants = { path = "../../../../../../../polkadot/runtime/rococo/constants", default-features = false }
 rococo-runtime = { path = "../../../../../../../polkadot/runtime/rococo" }
 
 # Cumulus
 parachains-common = { path = "../../../../../../parachains/common" }
-xcm-emulator = { path = "../../../../../../xcm/xcm-emulator", default-features = false }
-integration-tests-common = { path = "../../../common", default-features = false }
+emulated-integration-tests-common = { path = "../../../common", default-features = false }
