+++ cumulus/parachains/integration-tests/emulated/chains/relays/wococo/src/lib.rs
@@ -16,7 +16,7 @@
 pub mod genesis;
 
 // Cumulus
-use integration_tests_common::{
+use emulated_integration_tests_common::{
     xcm_emulator::decl_test_relay_chains,
     impl_accounts_helpers_for_relay_chain,
     impl_assert_events_helpers_for_relay_chain,
@@ -24,7 +24,7 @@ use integration_tests_common::{
 	impl_send_transact_helpers_for_relay_chain,
 };
 
-// Rococo declaration
+// Wococo declaration
 decl_test_relay_chains! {
 	#[api_version(8)]
 	pub struct Wococo {
@@ -43,7 +43,7 @@ decl_test_relay_chains! {
 	},
 }
 
-// Rococo implementation
+// Wococo implementation
 impl_accounts_helpers_for_relay_chain!(Wococo);
 impl_assert_events_helpers_for_relay_chain!(Wococo);
 impl_hrmp_channels_helpers_for_relay_chain!(Wococo);
