+++ cumulus/parachains/integration-tests/emulated/common/src/lib.rs
@@ -13,335 +13,150 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod constants;
 pub mod impls;
 pub mod macros;
 pub mod xcm_helpers;
 
 pub use xcm_emulator;
 
-// use constants::{
-// 	accounts::{ALICE, BOB},
-// 	asset_hub_rococo, asset_hub_westend, asset_hub_wococo, bridge_hub_rococo, penpal, rococo,
-// 	westend,
-// };
-// use impls::{RococoWococoMessageHandler, WococoRococoMessageHandler};
-// pub use paste;
+// Substrate
+use grandpa::AuthorityId as GrandpaId;
+use pallet_im_online::sr25519::AuthorityId as ImOnlineId;
+use sp_authority_discovery::AuthorityId as AuthorityDiscoveryId;
+use sp_consensus_babe::AuthorityId as BabeId;
+use sp_core::{sr25519, Pair, Public};
+use sp_runtime::{
+	traits::{IdentifyAccount, Verify},
+	MultiSignature,
+};
 
-// // Substrate
-// use frame_support::traits::OnInitialize;
-// pub use pallet_balances;
+// Polakdot
+use polkadot_runtime_parachains::configuration::HostConfiguration;
+use parachains_common::BlockNumber;
+use xcm;
 
-// // Cumulus
-// pub use cumulus_pallet_xcmp_queue;
-// pub use xcm_emulator::Chain;
-// use xcm_emulator::{
-// 	decl_test_bridges, decl_test_networks, decl_test_parachains, decl_test_relay_chains,
-// 	decl_test_sender_receiver_accounts_parameter_types,
-// };
+// Cumulus
+use parachains_common::{AccountId, AssetHubPolkadotAuraId, AuraId};
+use polkadot_primitives::{AssignmentId, ValidatorId};
+use polkadot_service::chain_spec::get_authority_keys_from_seed_no_beefy;
 
-// // Polkadot
-// pub use pallet_xcm;
-// pub use xcm::prelude::{AccountId32, WeightLimit};
+pub const XCM_V2: u32 = 2;
+pub const XCM_V3: u32 = 3;
+pub const REF_TIME_THRESHOLD: u64 = 33;
+pub const PROOF_SIZE_THRESHOLD: u64 = 33;
 
-// decl_test_relay_chains! {
-// 	#[api_version(8)]
-// 	pub struct Westend {
-// 		genesis = westend::genesis(),
-// 		on_init = (),
-// 		runtime = westend_runtime,
-// 		core = {
-// 			SovereignAccountOf: westend_runtime::xcm_config::LocationConverter, //TODO: rename to SovereignAccountOf,
-// 		},
-// 		pallets = {
-// 			XcmPallet: westend_runtime::XcmPallet,
-// 			Sudo: westend_runtime::Sudo,
-// 			Balances: westend_runtime::Balances,
-// 			Treasury: westend_runtime::Treasury,
-// 			AssetRate: westend_runtime::AssetRate,
-// 		}
-// 	},
-// 	#[api_version(8)]
-// 	pub struct Rococo {
-// 		genesis = rococo::genesis(),
-// 		on_init = (),
-// 		runtime = rococo_runtime,
-// 		core = {
-// 			SovereignAccountOf: rococo_runtime::xcm_config::LocationConverter, //TODO: rename to SovereignAccountOf,
-// 		},
-// 		pallets = {
-// 			XcmPallet: rococo_runtime::XcmPallet,
-// 			Sudo: rococo_runtime::Sudo,
-// 			Balances: rococo_runtime::Balances,
-// 			Hrmp: rococo_runtime::Hrmp,
-// 		}
-// 	},
-// 	#[api_version(8)]
-// 	pub struct Wococo {
-// 		genesis = rococo::genesis(),
-// 		on_init = (),
-// 		runtime = rococo_runtime,
-// 		core = {
-// 			SovereignAccountOf: rococo_runtime::xcm_config::LocationConverter, //TODO: rename to SovereignAccountOf,
-// 		},
-// 		pallets = {
-// 			XcmPallet: rococo_runtime::XcmPallet,
-// 			Sudo: rococo_runtime::Sudo,
-// 			Balances: rococo_runtime::Balances,
-// 		}
-// 	}
-// }
+/// The default XCM version to set in genesis config.
+pub const SAFE_XCM_VERSION: u32 = xcm::prelude::XCM_VERSION;
 
-// decl_test_parachains! {
-// 	// Westend Parachains
-// 	pub struct AssetHubWestend {
-// 		genesis = asset_hub_westend::genesis(),
-// 		on_init = {
-// 			asset_hub_westend_runtime::AuraExt::on_initialize(1);
-// 		},
-// 		runtime = asset_hub_westend_runtime,
-// 		core = {
-// 			XcmpMessageHandler: asset_hub_westend_runtime::XcmpQueue,
-// 			DmpMessageHandler: asset_hub_westend_runtime::DmpQueue,
-// 			LocationToAccountId: asset_hub_westend_runtime::xcm_config::LocationToAccountId,
-// 			ParachainInfo: asset_hub_westend_runtime::ParachainInfo,
-// 		},
-// 		pallets = {
-// 			PolkadotXcm: asset_hub_westend_runtime::PolkadotXcm,
-// 			Balances: asset_hub_westend_runtime::Balances,
-// 			Assets: asset_hub_westend_runtime::Assets,
-// 			ForeignAssets: asset_hub_westend_runtime::ForeignAssets,
-// 			PoolAssets: asset_hub_westend_runtime::PoolAssets,
-// 			AssetConversion: asset_hub_westend_runtime::AssetConversion,
-// 		}
-// 	},
-// 	pub struct PenpalA {
-// 		genesis = penpal::genesis(penpal::PARA_ID_A),
-// 		on_init = {
-// 			penpal_runtime::AuraExt::on_initialize(1);
-// 		},
-// 		runtime = penpal_runtime,
-// 		core = {
-// 			XcmpMessageHandler: penpal_runtime::XcmpQueue,
-// 			DmpMessageHandler: penpal_runtime::DmpQueue,
-// 			LocationToAccountId: penpal_runtime::xcm_config::LocationToAccountId,
-// 			ParachainInfo: penpal_runtime::ParachainInfo,
-// 		},
-// 		pallets = {
-// 			PolkadotXcm: penpal_runtime::PolkadotXcm,
-// 			Assets: penpal_runtime::Assets,
-// 			Balances: penpal_runtime::Balances,
-// 		}
-// 	},
-// 	// Rococo Parachains
-// 	pub struct BridgeHubRococo {
-// 		genesis = bridge_hub_rococo::genesis(),
-// 		on_init = {
-// 			bridge_hub_rococo_runtime::AuraExt::on_initialize(1);
-// 		},
-// 		runtime = bridge_hub_rococo_runtime,
-// 		core = {
-// 			XcmpMessageHandler: bridge_hub_rococo_runtime::XcmpQueue,
-// 			DmpMessageHandler: bridge_hub_rococo_runtime::DmpQueue,
-// 			LocationToAccountId: bridge_hub_rococo_runtime::xcm_config::LocationToAccountId,
-// 			ParachainInfo: bridge_hub_rococo_runtime::ParachainInfo,
-// 		},
-// 		pallets = {
-// 			PolkadotXcm: bridge_hub_rococo_runtime::PolkadotXcm,
-// 			Balances: bridge_hub_rococo_runtime::Balances,
-// 		}
-// 	},
-// 	// AssetHubRococo
-// 	pub struct AssetHubRococo {
-// 		genesis = asset_hub_rococo::genesis(),
-// 		on_init = {
-// 			asset_hub_rococo_runtime::AuraExt::on_initialize(1);
-// 		},
-// 		runtime = asset_hub_rococo_runtime,
-// 		core = {
-// 			XcmpMessageHandler: asset_hub_rococo_runtime::XcmpQueue,
-// 			DmpMessageHandler: asset_hub_rococo_runtime::DmpQueue,
-// 			LocationToAccountId: asset_hub_rococo_runtime::xcm_config::LocationToAccountId,
-// 			ParachainInfo: asset_hub_rococo_runtime::ParachainInfo,
-// 		},
-// 		pallets = {
-// 			PolkadotXcm: asset_hub_rococo_runtime::PolkadotXcm,
-// 			Assets: asset_hub_rococo_runtime::Assets,
-// 			ForeignAssets: asset_hub_rococo_runtime::ForeignAssets,
-// 			PoolAssets: asset_hub_rococo_runtime::PoolAssets,
-// 			AssetConversion: asset_hub_rococo_runtime::AssetConversion,
-// 			Balances: asset_hub_rococo_runtime::Balances,
-// 		}
-// 	},
-// 	pub struct PenpalA {
-// 		genesis = penpal::genesis(penpal::PARA_ID_A),
-// 		on_init = {
-// 			penpal_runtime::AuraExt::on_initialize(1);
-// 		},
-// 		runtime = penpal_runtime,
-// 		core = {
-// 			XcmpMessageHandler: penpal_runtime::XcmpQueue,
-// 			DmpMessageHandler: penpal_runtime::DmpQueue,
-// 			LocationToAccountId: penpal_runtime::xcm_config::LocationToAccountId,
-// 			ParachainInfo: penpal_runtime::ParachainInfo,
-// 		},
-// 		pallets = {
-// 			PolkadotXcm: penpal_runtime::PolkadotXcm,
-// 			Assets: penpal_runtime::Assets,
-// 		}
-// 	},
-// 	pub struct PenpalRococoB {
-// 		genesis = penpal::genesis(penpal::PARA_ID_B),
-// 		on_init = {
-// 			penpal_runtime::AuraExt::on_initialize(1);
-// 		},
-// 		runtime = penpal_runtime,
-// 		core = {
-// 			XcmpMessageHandler: penpal_runtime::XcmpQueue,
-// 			DmpMessageHandler: penpal_runtime::DmpQueue,
-// 			LocationToAccountId: penpal_runtime::xcm_config::LocationToAccountId,
-// 			ParachainInfo: penpal_runtime::ParachainInfo,
-// 		},
-// 		pallets = {
-// 			PolkadotXcm: penpal_runtime::PolkadotXcm,
-// 			Assets: penpal_runtime::Assets,
-// 		}
-// 	},
-// 	// Wococo Parachains
-// 	pub struct BridgeHubWococo {
-// 		genesis = bridge_hub_rococo::genesis(),
-// 		on_init = {
-// 			bridge_hub_rococo_runtime::AuraExt::on_initialize(1);
-// 			// TODO: manage to set_wococo_flavor with `set_storage`
-// 		},
-// 		runtime = bridge_hub_rococo_runtime,
-// 		core = {
-// 			XcmpMessageHandler: bridge_hub_rococo_runtime::XcmpQueue,
-// 			DmpMessageHandler: bridge_hub_rococo_runtime::DmpQueue,
-// 			LocationToAccountId: bridge_hub_rococo_runtime::xcm_config::LocationToAccountId,
-// 			ParachainInfo: bridge_hub_rococo_runtime::ParachainInfo,
-// 		},
-// 		pallets = {
-// 			PolkadotXcm: bridge_hub_rococo_runtime::PolkadotXcm,
-// 		}
-// 	},
-// 	pub struct AssetHubWococo {
-// 		genesis = asset_hub_wococo::genesis(),
-// 		on_init = {
-// 			asset_hub_rococo_runtime::AuraExt::on_initialize(1);
-// 			// TODO: manage to set_wococo_flavor with `set_storage`
-// 		},
-// 		runtime = asset_hub_rococo_runtime,
-// 		core = {
-// 			XcmpMessageHandler: asset_hub_rococo_runtime::XcmpQueue,
-// 			DmpMessageHandler: asset_hub_rococo_runtime::DmpQueue,
-// 			LocationToAccountId: asset_hub_rococo_runtime::xcm_config::LocationToAccountId,
-// 			ParachainInfo: asset_hub_rococo_runtime::ParachainInfo,
-// 		},
-// 		pallets = {
-// 			PolkadotXcm: asset_hub_rococo_runtime::PolkadotXcm,
-// 			Assets: asset_hub_rococo_runtime::Assets,
-// 			ForeignAssets: asset_hub_rococo_runtime::ForeignAssets,
-// 			PoolAssets: asset_hub_rococo_runtime::PoolAssets,
-// 			AssetConversion: asset_hub_rococo_runtime::AssetConversion,
-// 			Balances: asset_hub_rococo_runtime::Balances,
-// 		}
-// 	}
-// }
+type AccountPublic = <MultiSignature as Verify>::Signer;
 
-// decl_test_networks! {
-// 	pub struct WestendMockNet {
-// 		relay_chain = Westend,
-// 		parachains = vec![
-// 			AssetHubWestend,
-// 			PenpalA,
-// 		],
-// 		bridge = ()
-// 	},
-// 	pub struct RococoMockNet {
-// 		relay_chain = Rococo,
-// 		parachains = vec![
-// 			AssetHubRococo,
-// 			BridgeHubRococo,
-// 			PenpalA,
-// 			PenpalRococoB,
-// 		],
-// 		bridge = RococoWococoMockBridge
-// 	},
-// 	pub struct WococoMockNet {
-// 		relay_chain = Wococo,
-// 		parachains = vec![
-// 			AssetHubWococo,
-// 			BridgeHubWococo,
-// 		],
-// 		bridge = WococoRococoMockBridge
-// 	}
-// }
+/// Helper function to generate a crypto pair from seed
+pub fn get_from_seed<TPublic: Public>(seed: &str) -> <TPublic::Pair as Pair>::Public {
+	TPublic::Pair::from_string(&format!("//{}", seed), None)
+		.expect("static values are valid; qed")
+		.public()
+}
 
-// decl_test_bridges! {
-// 	pub struct RococoWococoMockBridge {
-// 		source = BridgeHubRococoPara,
-// 		target = BridgeHubWococoPara,
-// 		handler = RococoWococoMessageHandler
-// 	},
-// 	pub struct WococoRococoMockBridge {
-// 		source = BridgeHubWococoPara,
-// 		target = BridgeHubRococoPara,
-// 		handler = WococoRococoMessageHandler
-// 	}
-// }
+/// Helper function to generate an account ID from seed.
+pub fn get_account_id_from_seed<TPublic: Public>(seed: &str) -> AccountId
+where
+	AccountPublic: From<<TPublic::Pair as Pair>::Public>,
+{
+	AccountPublic::from(get_from_seed::<TPublic>(seed)).into_account()
+}
 
-// // Westend implementation
-// impl_accounts_helpers_for_relay_chain!(Westend);
-// impl_assert_events_helpers_for_relay_chain!(Westend);
-// impl_send_transact_helpers_for_relay_chain!(Westend);
+pub fn get_host_config() -> HostConfiguration<BlockNumber> {
+	HostConfiguration {
+		max_upward_queue_count: 10,
+		max_upward_queue_size: 51200,
+		max_upward_message_size: 51200,
+		max_upward_message_num_per_candidate: 10,
+		max_downward_message_size: 51200,
+		hrmp_sender_deposit: 0,
+		hrmp_recipient_deposit: 0,
+		hrmp_channel_max_capacity: 1000,
+		hrmp_channel_max_message_size: 102400,
+		hrmp_channel_max_total_size: 102400,
+		hrmp_max_parachain_outbound_channels: 30,
+		hrmp_max_parachain_inbound_channels: 30,
+		..Default::default()
+	}
+}
 
-// // Rococo implementation
-// impl_accounts_helpers_for_relay_chain!(Rococo);
-// impl_assert_events_helpers_for_relay_chain!(Rococo);
-// impl_hrmp_channels_helpers_for_relay_chain!(Rococo);
-// impl_send_transact_helpers_for_relay_chain!(Rococo);
+pub mod accounts {
+	use super::*;
+	pub const ALICE: &str = "Alice";
+	pub const BOB: &str = "Bob";
+	pub const CHARLIE: &str = "Charlie";
+	pub const DAVE: &str = "Dave";
+	pub const EVE: &str = "Eve";
+	pub const FERDIE: &str = "Ferdei";
+	pub const ALICE_STASH: &str = "Alice//stash";
+	pub const BOB_STASH: &str = "Bob//stash";
+	pub const CHARLIE_STASH: &str = "Charlie//stash";
+	pub const DAVE_STASH: &str = "Dave//stash";
+	pub const EVE_STASH: &str = "Eve//stash";
+	pub const FERDIE_STASH: &str = "Ferdie//stash";
+	pub const FERDIE_BEEFY: &str = "Ferdie//stash";
 
-// // Wococo implementation
-// impl_accounts_helpers_for_relay_chain!(Wococo);
-// impl_assert_events_helpers_for_relay_chain!(Wococo);
-// impl_send_transact_helpers_for_relay_chain!(Wococo);
+	pub fn init_balances() -> Vec<AccountId> {
+		vec![
+			get_account_id_from_seed::<sr25519::Public>(ALICE),
+			get_account_id_from_seed::<sr25519::Public>(BOB),
+			get_account_id_from_seed::<sr25519::Public>(CHARLIE),
+			get_account_id_from_seed::<sr25519::Public>(DAVE),
+			get_account_id_from_seed::<sr25519::Public>(EVE),
+			get_account_id_from_seed::<sr25519::Public>(FERDIE),
+			get_account_id_from_seed::<sr25519::Public>(ALICE_STASH),
+			get_account_id_from_seed::<sr25519::Public>(BOB_STASH),
+			get_account_id_from_seed::<sr25519::Public>(CHARLIE_STASH),
+			get_account_id_from_seed::<sr25519::Public>(DAVE_STASH),
+			get_account_id_from_seed::<sr25519::Public>(EVE_STASH),
+			get_account_id_from_seed::<sr25519::Public>(FERDIE_STASH),
+		]
+	}
+}
 
-// // AssetHubWestend implementation
-// impl_accounts_helpers_for_parachain!(AssetHubWestend);
-// impl_assets_helpers_for_system_parachain!(AssetHubWestend, Westend);
-// impl_assert_events_helpers_for_parachain!(AssetHubWestend);
+pub mod collators {
+	use super::*;
 
-// // AssetHubRococo implementation
-// impl_accounts_helpers_for_parachain!(AssetHubRococo);
-// impl_assets_helpers_for_system_parachain!(AssetHubRococo, Rococo);
-// impl_assert_events_helpers_for_parachain!(AssetHubRococo);
+	pub fn invulnerables_asset_hub_polkadot() -> Vec<(AccountId, AssetHubPolkadotAuraId)> {
+		vec![
+			(
+				get_account_id_from_seed::<sr25519::Public>("Alice"),
+				get_from_seed::<AssetHubPolkadotAuraId>("Alice"),
+			),
+			(
+				get_account_id_from_seed::<sr25519::Public>("Bob"),
+				get_from_seed::<AssetHubPolkadotAuraId>("Bob"),
+			),
+		]
+	}
 
-// // PenpalA implementation
-// impl_assert_events_helpers_for_parachain!(PenpalA);
+	pub fn invulnerables() -> Vec<(AccountId, AuraId)> {
+		vec![
+			(
+				get_account_id_from_seed::<sr25519::Public>("Alice"),
+				get_from_seed::<AuraId>("Alice"),
+			),
+			(get_account_id_from_seed::<sr25519::Public>("Bob"), get_from_seed::<AuraId>("Bob")),
+		]
+	}
+}
 
-// // BridgeHubRococo implementation
-// impl_accounts_helpers_for_parachain!(BridgeHubRococo);
-// impl_assert_events_helpers_for_parachain!(BridgeHubRococo);
+pub mod validators {
+	use super::*;
 
-// // PenpalRococo implementations
-// impl_assert_events_helpers_for_parachain!(PenpalA);
-// impl_assert_events_helpers_for_parachain!(PenpalRococoB);
-
-// decl_test_sender_receiver_accounts_parameter_types! {
-// 	// Relays
-// 	WestendRelay { sender: ALICE, receiver: BOB },
-// 	RococoRelay { sender: ALICE, receiver: BOB },
-// 	WococoRelay { sender: ALICE, receiver: BOB },
-// 	// Asset Hubs
-// 	AssetHubWestendPara { sender: ALICE, receiver: BOB },
-// 	AssetHubRococoPara { sender: ALICE, receiver: BOB },
-// 	AssetHubWococoPara { sender: ALICE, receiver: BOB },
-// 	// Bridged Hubs
-// 	BridgeHubRococoPara { sender: ALICE, receiver: BOB },
-// 	BridgeHubWococoPara { sender: ALICE, receiver: BOB },
-// 	// Penpals
-// 	PenpalAPara { sender: ALICE, receiver: BOB },
-// 	PenpalAPara { sender: ALICE, receiver: BOB },
-// 	PenpalRococoBPara { sender: ALICE, receiver: BOB }
-// }
+	pub fn initial_authorities() -> Vec<(
+		AccountId,
+		AccountId,
+		BabeId,
+		GrandpaId,
+		ImOnlineId,
+		ValidatorId,
+		AssignmentId,
+		AuthorityDiscoveryId,
+	)> {
+		vec![get_authority_keys_from_seed_no_beefy("Alice")]
+	}
+}
