+++ cumulus/parachains/integration-tests/emulated/common/src/macros.rs
@@ -14,11 +14,17 @@
 // limitations under the License.
 
 pub use paste;
-pub use xcm_emulator::Chain;
-pub use xcm::prelude::{AccountId32, WeightLimit};
+
+// Substrate
 pub use pallet_xcm;
-pub use cumulus_pallet_xcmp_queue;
 pub use pallet_balances;
+
+// Polkadot
+pub use xcm::prelude::{AccountId32, WeightLimit};
+
+// Cumulus
+pub use cumulus_pallet_xcmp_queue;
+pub use xcm_emulator::Chain;
 
 #[macro_export]
 macro_rules! test_parachain_is_trusted_teleporter {
