+++ cumulus/parachains/integration-tests/emulated/common/src/xcm_helpers.rs
@@ -13,7 +13,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// Cumulus
 use parachains_common::AccountId;
+
+// Polkadot
 use xcm::{prelude::*, DoubleEncoded};
 
 /// Helper method to build a XCM with a `Transact` instruction and paying for its execution
