+++ cumulus/parachains/integration-tests/emulated/networks/rococo-system/Cargo.toml
@@ -9,10 +9,8 @@ publish = false
 
 [dependencies]
 # Cumulus
-xcm-emulator = { path = "../../../../../xcm/xcm-emulator", default-features = false }
-integration-tests-common = { path = "../../common", default-features = false }
+emulated-integration-tests-common = { path = "../../common", default-features = false }
 rococo-emulated-chain = { path = "../../chains/relays/rococo" }
 asset-hub-rococo-emulated-chain = { path = "../../chains/parachains/assets/asset-hub-rococo" }
 bridge-hub-rococo-emulated-chain = { path = "../../chains/parachains/bridges/bridge-hub-rococo" }
 penpal-emulated-chain = { path = "../../chains/parachains/testing/penpal" }
-# wococo-system-emulated-network = { path = "../wococo-system" }
