+++ cumulus/parachains/integration-tests/emulated/networks/rococo-system/src/lib.rs
@@ -13,7 +13,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Local
 pub use rococo_emulated_chain;
 pub use asset_hub_rococo_emulated_chain;
 pub use bridge_hub_rococo_emulated_chain;
@@ -25,8 +24,8 @@ use bridge_hub_rococo_emulated_chain::BridgeHubRococo;
 use penpal_emulated_chain::{PenpalA, PenpalB};
 
 // Cumulus
-use integration_tests_common::{
-    constants::accounts::{ALICE, BOB},
+use emulated_integration_tests_common::{
+    accounts::{ALICE, BOB},
     xcm_emulator::{
         decl_test_networks,
         decl_test_sender_receiver_accounts_parameter_types,
@@ -42,7 +41,6 @@ decl_test_networks! {
 			PenpalA,
 			PenpalB,
 		],
-		// bridge = RococoWococoMockBridge
 		bridge = ()
 	},
 }
