+++ cumulus/parachains/integration-tests/emulated/networks/westend-system/src/lib.rs
@@ -13,7 +13,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Local
 pub use westend_emulated_chain;
 pub use asset_hub_westend_emulated_chain;
 pub use penpal_emulated_chain;
@@ -23,8 +22,8 @@ use asset_hub_westend_emulated_chain::AssetHubWestend;
 use penpal_emulated_chain::{PenpalA, PenpalB};
 
 // Cumulus
-use integration_tests_common::{
-	constants::accounts::{ALICE, BOB},
+use emulated_integration_tests_common::{
+	accounts::{ALICE, BOB},
     xcm_emulator::{
         decl_test_networks,
         decl_test_sender_receiver_accounts_parameter_types,
