+++ cumulus/parachains/integration-tests/emulated/tests/assets/asset-hub-rococo/Cargo.toml
@@ -14,15 +14,11 @@ assert_matches = "1.5.0"
 # Substrate
 sp-runtime = { path = "../../../../../../../substrate/primitives/runtime", default-features = false}
 frame-support = { path = "../../../../../../../substrate/frame/support", default-features = false}
-frame-system = { path = "../../../../../../../substrate/frame/system", default-features = false}
 pallet-balances = { path = "../../../../../../../substrate/frame/balances", default-features = false}
 pallet-assets = { path = "../../../../../../../substrate/frame/assets", default-features = false}
 pallet-asset-conversion = { path = "../../../../../../../substrate/frame/asset-conversion", default-features = false}
 
 # Polkadot
-polkadot-core-primitives = { path = "../../../../../../../polkadot/core-primitives", default-features = false}
-polkadot-parachain-primitives = { path = "../../../../../../../polkadot/parachain", default-features = false}
-polkadot-runtime-parachains = { path = "../../../../../../../polkadot/runtime/parachains" }
 xcm = { package = "staging-xcm", path = "../../../../../../../polkadot/xcm", default-features = false}
 pallet-xcm = { path = "../../../../../../../polkadot/xcm/pallet-xcm", default-features = false}
 xcm-executor = { package = "staging-xcm-executor",  path = "../../../../../../../polkadot/xcm/xcm-executor", default-features = false}
@@ -32,26 +28,5 @@ rococo-runtime = { path = "../../../../../../../polkadot/runtime/rococo" }
 asset-test-utils = { path = "../../../../../runtimes/assets/test-utils", default-features = false }
 parachains-common = { path = "../../../../../../parachains/common" }
 asset-hub-rococo-runtime = { path = "../../../../../runtimes/assets/asset-hub-rococo" }
-
-# Local
-xcm-emulator = { path = "../../../../../../xcm/xcm-emulator", default-features = false}
-integration-tests-common = { path = "../../../common", default-features = false}
+emulated-integration-tests-common = { path = "../../../common", default-features = false}
 rococo-system-emulated-network ={ path = "../../../networks/rococo-system" }
-
-# [features]
-# runtime-benchmarks = [
-# 	"asset-hub-rococo-runtime/runtime-benchmarks",
-# 	"frame-support/runtime-benchmarks",
-# 	"frame-system/runtime-benchmarks",
-# 	"integration-tests-common/runtime-benchmarks",
-# 	"pallet-asset-conversion/runtime-benchmarks",
-# 	"pallet-assets/runtime-benchmarks",
-# 	"pallet-balances/runtime-benchmarks",
-# 	"pallet-xcm/runtime-benchmarks",
-# 	"parachains-common/runtime-benchmarks",
-# 	"polkadot-parachain-primitives/runtime-benchmarks",
-# 	"polkadot-runtime-parachains/runtime-benchmarks",
-# 	"rococo-runtime/runtime-benchmarks",
-# 	"sp-runtime/runtime-benchmarks",
-# 	"xcm-executor/runtime-benchmarks",
-# ]
