+++ cumulus/parachains/integration-tests/emulated/tests/assets/asset-hub-westend/src/lib.rs
@@ -13,8 +13,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub use asset_test_utils::xcm_helpers;
 pub use codec::Encode;
+
+// Substrate
 pub use frame_support::{
 	assert_err, assert_ok,
 	instances::Instance2,
@@ -24,15 +25,14 @@ pub use frame_support::{
 	BoundedVec,
 };
 
-pub use integration_tests_common::{
-	constants::{
-		PROOF_SIZE_THRESHOLD,
-		REF_TIME_THRESHOLD, XCM_V3,
-	},
-	test_parachain_is_trusted_teleporter,
-	xcm_helpers::{xcm_transact_paid_execution, xcm_transact_unpaid_execution},
+// Polkadot
+pub use xcm::{
+	prelude::{AccountId32 as AccountId32Junction, *},
+	v3::{Error, NetworkId::Westend as WestendId},
 };
 
+// Cumulus
+pub use parachains_common::{AccountId, Balance};
 pub use westend_system_emulated_network::{
 	westend_emulated_chain::{
 		genesis::ED as WESTEND_ED,
@@ -50,16 +50,17 @@ pub use westend_system_emulated_network::{
 	PenpalAParaReceiver as PenpalAReceiver, PenpalAParaSender as PenpalASender,
 	PenpalAPara as PenpalA,
 };
-pub use parachains_common::{AccountId, Balance};
-pub use xcm::{
-	prelude::{AccountId32 as AccountId32Junction, *},
-	v3::{Error, NetworkId::Westend as WestendId},
+pub use emulated_integration_tests_common::{
+	PROOF_SIZE_THRESHOLD,
+	REF_TIME_THRESHOLD, XCM_V3,
+	test_parachain_is_trusted_teleporter,
+	xcm_helpers::{xcm_transact_paid_execution, xcm_transact_unpaid_execution},
+	xcm_emulator::{
+		assert_expected_events, bx, helpers::weight_within_threshold, Chain, Parachain as Para,
+		RelayChain as Relay, Test, TestArgs, TestContext, TestExt,
+	}
 };
-
-pub use xcm_emulator::{
-	assert_expected_events, bx, helpers::weight_within_threshold, Chain, Parachain as Para,
-	RelayChain as Relay, Test, TestArgs, TestContext, TestExt,
-};
+pub use asset_test_utils::xcm_helpers;
 
 pub const ASSET_ID: u32 = 1;
 pub const ASSET_MIN_BALANCE: u128 = 1000;
