+++ cumulus/parachains/integration-tests/emulated/tests/bridges/bridge-hub-rococo/Cargo.toml
@@ -14,9 +14,6 @@ codec = { package = "parity-scale-codec", version = "3.4.0", default-features = 
 frame-support = { path = "../../../../../../../substrate/frame/support", default-features = false}
 
 # Polkadot
-polkadot-core-primitives = { path = "../../../../../../../polkadot/core-primitives", default-features = false}
-polkadot-parachain-primitives = { path = "../../../../../../../polkadot/parachain", default-features = false}
-polkadot-runtime-parachains = { path = "../../../../../../../polkadot/runtime/parachains" }
 xcm = { package = "staging-xcm", path = "../../../../../../../polkadot/xcm", default-features = false}
 pallet-xcm = { path = "../../../../../../../polkadot/xcm/pallet-xcm", default-features = false}
 xcm-executor = { package = "staging-xcm-executor",  path = "../../../../../../../polkadot/xcm/xcm-executor", default-features = false}
@@ -31,8 +28,5 @@ parachains-common = { path = "../../../../../../parachains/common" }
 cumulus-pallet-xcmp-queue = { path = "../../../../../../pallets/xcmp-queue", default-features = false}
 cumulus-pallet-dmp-queue = { path = "../../../../../../pallets/dmp-queue", default-features = false}
 bridge-hub-rococo-runtime = { path = "../../../../../../parachains/runtimes/bridge-hubs/bridge-hub-rococo", default-features = false }
-
-# Local
-xcm-emulator = { path = "../../../../../../xcm/xcm-emulator", default-features = false}
-integration-tests-common = { path = "../../../common", default-features = false}
+emulated-integration-tests-common = { path = "../../../common", default-features = false}
 rococo-wococo-system-emulated-network ={ path = "../../../networks/rococo-wococo-system" }
