+++ cumulus/parachains/integration-tests/emulated/tests/bridges/bridge-hub-rococo/src/lib.rs
@@ -13,16 +13,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub use bp_messages::LaneId;
+// Substrate
 pub use frame_support::assert_ok;
-pub use integration_tests_common::{
-	constants::{
-		PROOF_SIZE_THRESHOLD,
-		REF_TIME_THRESHOLD, XCM_V3,
+
+// Polkadot
+pub use xcm::{
+	prelude::{AccountId32 as AccountId32Junction, *},
+	v3::{
+		Error,
+		NetworkId::{Rococo as RococoId, Wococo as WococoId},
 	},
-	test_parachain_is_trusted_teleporter,
-	xcm_helpers::{xcm_transact_paid_execution, xcm_transact_unpaid_execution},
 };
+
+// Bridges
+pub use bp_messages::LaneId;
+
+// Cumulus
+pub use parachains_common::{AccountId, Balance};
 pub use rococo_wococo_system_emulated_network::{
 	rococo_emulated_chain::{
 		genesis::ED as ROCOCO_ED,
@@ -40,25 +47,16 @@ pub use rococo_wococo_system_emulated_network::{
 	BridgeHubRococoParaSender as BridgeHubRococoSender,
 	AssetHubWococoPara as AssetHubWococo,
 	BridgeHubWococoPara as BridgeHubWococo,
-	// BridgeHubRococoParaPallet as BridgeHubWococoPallet,
 };
-// pub use wococo_system_emulated_network::{
-// 	bridge_hub_rococo_emulated_chain::genesis::ED as BRIDGE_HUB_WOCOCO_ED,
-// 	AssetHubWococoPara as AssetHubWococo,
-// 	BridgeHubWococoPara as BridgeHubWococo,
-// 	BridgeHubWococoParaPallet as BridgeHubWococoPallet,
-// };
-pub use parachains_common::{AccountId, Balance};
-pub use xcm::{
-	prelude::{AccountId32 as AccountId32Junction, *},
-	v3::{
-		Error,
-		NetworkId::{Rococo as RococoId, Wococo as WococoId},
-	},
-};
-pub use xcm_emulator::{
-	assert_expected_events, bx, helpers::weight_within_threshold, Chain, Parachain as Para,
-	RelayChain as Relay, Test, TestArgs, TestContext, TestExt,
+pub use emulated_integration_tests_common::{
+	PROOF_SIZE_THRESHOLD,
+	REF_TIME_THRESHOLD, XCM_V3,
+	test_parachain_is_trusted_teleporter,
+	xcm_helpers::{xcm_transact_paid_execution, xcm_transact_unpaid_execution},
+	xcm_emulator::{
+		assert_expected_events, bx, helpers::weight_within_threshold, Chain, Parachain as Para,
+		RelayChain as Relay, Test, TestArgs, TestContext, TestExt,
+	}
 };
 
 pub const ASSET_ID: u32 = 1;
